book_embeddings/requirements.txt
langchain==0.1.16
openai==1.14.2
python-dotenv==1.0.0
PyPDF2==3.0.1
tiktoken==0.6.0
chromadb==0.4.24
sentence-transformers==2.3.0